requirements.txt
Django==6.0.1
djangorestframework==3.16.1
django-cors-headers==4.9.0
psycopg2-binary==2.9.11
celery==5.6.2
redis==7.1.0
django-otp==1.7.0
django-ratelimit==4.1.0
django-auditlog==3.4.1
gunicorn==23.0.0
whitenoise==6.8.2
beautifulsoup4==4.12.3
requests==2.31.0
lxml==5.1.0
channels==4.0.0
channels-redis==4.2.0
Pillow==10.4.0
django-phonenumber-field==7.2.0
phonenumbers==8.13.36
